move endpoint helper to util
