+++ util_test.go
@@ -5,6 +5,12 @@ import (
 	"time"
 )
 
+func TestSpotEP(t *testing.T) {
+	if spotEP(c, "123") != "http://magicseaweed.com/api/fakeKey/forecast/?spot_id=123" {
+		t.Error("spotEp should return the proper endpoint")
+	}
+}
+
 func TestConcat(t *testing.T) {
 	joined := concat([]string{
 		"foo",
